requirements.txt
google-generativeai 
google-auth
fastapi 
uvicorn 
snowflake-connector-python 
google-generativeai 
python-multipart
python-dotenv
Jinja2

